requirements.txt
# Web frameworks and utilities
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
python-dotenv==1.0.0
jinja2==3.1.2
pydantic==2.4.2

# Monitoring and metrics
prometheus-client==0.19.0
psutil==5.9.6

# Scientific computing (for simulator)
numpy==1.26.1

# CORS support
starlette==0.27.0

# Template rendering
aiofiles==23.2.1

# Type hints
typing-extensions==4.8.0

# OpenAI (for chatbot)
openai==1.3.0

# HTTP server
gunicorn==21.2.0
